Reduce variables to uint32

+++ contracts/Utils.sol
@@ -5,26 +5,26 @@ import "./Errors.sol";
 
 library Pagination {
     function getPagination(
-        uint256 _length,
-        uint256 _pageSize,
-        uint256 _pageNumber
+        uint32 _length,
+        uint32 _pageSize,
+        uint32 _pageNumber
     )
         internal
         pure
         returns (
-            uint256,
-            uint256,
-            uint256
+            uint32,
+            uint32,
+            uint32
         )
     {
-        uint256 pages = _length / _pageSize;
-        uint256 lastPage = _length - pages * _pageSize > 0 ? 1 : 0;
-        uint256 pagesCount = pages + lastPage;
+        uint32 pages = _length / _pageSize;
+        uint32 lastPage = _length - pages * _pageSize > 0 ? 1 : 0;
+        uint32 pagesCount = pages + lastPage;
         if (_pageNumber > pagesCount) revert PageOutOfRange();
-        uint256 end = _pageNumber * (_pageSize) < _length
+        uint32 end = _pageNumber * (_pageSize) < _length
             ? _pageNumber * (_pageSize)
             : _length;
-        uint256 start = _pageSize * (_pageNumber - 1);
+        uint32 start = _pageSize * (_pageNumber - 1);
         return (start, end, pagesCount);
     }
 }
